# auth_service/docs/swagger.yaml
basePath: /
...
host: localhost:8082
info:
  contact: {}
  description: Сервис авторизации
  title: Auth Service API
  version: "1.0"
paths:
...
  /account/delete/request-confirmation:
    post:
      description: |-
        Отправляет magic-link код на email текущего пользователя для
        подтверждения удаления аккаунта. Требуется только для
        oauth-only пользователей без пароля — у них нет иного
        способа подтвердить чувствительное действие. Возвращает
        session_id, который затем передаётся вместе с кодом из письма
        в /account/delete.
      produces:
      - application/json
      responses:
        "200":
          description: Код отправлен на email
          schema:
            properties:
              session_id:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Access token отсутствует, невалиден или истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Запросить подтверждение удаления аккаунта через magic link
      tags:
      - account
...
  /auth/2fa/disable/request-confirmation:
    post:
      description: |-
        Отправляет magic-link код на email текущего пользователя для
        подтверждения отключения 2FA. Требуется только для
        oauth-only пользователей без пароля — у них нет иного
        способа подтвердить чувствительное действие. Возвращает
        session_id, который затем передаётся вместе с кодом из письма
        в /auth/2fa/disable.
      produces:
      - application/json
      responses:
        "200":
          description: Код отправлен на email
          schema:
            properties:
              session_id:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Access token отсутствует, невалиден или истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Запросить подтверждение отключения 2FA через magic link
      tags:
      - 2fa
  /auth/2fa/magic-link/disable:
    post:
      consumes:
      - application/json
      description: |-
        Отключает magic-link 2FA. Подтверждение зависит от того, есть
        ли у пользователя пароль: если да — передаётся password; если
        нет (oauth-only аккаунт) — передаются session_id и token,
        полученные через /auth/2fa/magic-link/request-action-confirmation.
      parameters:
      - description: Подтверждение отключения (один из наборов полей)
        in: body
        name: request
        schema:
          properties:
            password:
              type: string
            session_id:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 2FA отключена
          schema:
            properties:
              status:
                type: string
            type: object
        "401":
          description: Access token отсутствует, невалиден или истёк, либо неверное
            подтверждение (пароль/magic-link код)
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "409":
          description: 2FA не включена
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отключить magic-link 2FA
      tags:
      - 2fa
  /auth/2fa/magic-link/enable:
    post:
      description: |-
        Включает magic-link 2FA для текущего пользователя. Требует,
        чтобы у пользователя уже был рабочий фактор для будущего
        отключения (пароль или хотя бы один привязанный oauth-аккаунт) —
        иначе включение необратимо заблокирует доступ к аккаунту.
      produces:
      - application/json
      responses:
        "200":
          description: 2FA включена
          schema:
            properties:
              status:
                type: string
            type: object
        "401":
          description: Access token отсутствует, невалиден или истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "409":
          description: 2FA уже включена, либо нет ни одного доступного фактора для
            будущего disable
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Включить magic-link 2FA
      tags:
      - 2fa
  /auth/2fa/magic-link/resend:
    post:
      consumes:
      - application/json
      description: |-
        Инвалидирует предыдущую активную ссылку и высылает новую в
        рамках той же pending-сессии, начатой на /auth/login. Не
        подтверждает и не раскрывает факт доставки письма — ответ
        одинаковый независимо от того, дошло письмо или нет.
      parameters:
      - description: Идентификатор pending-сессии
        in: body
        name: request
        required: true
        schema:
          properties:
            session_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Новая ссылка отправлена (либо попытка предпринята)
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Невалидное тело запроса
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Pending-сессия не найдена или истекла — нужно начать логин
            заново
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "429":
          description: Слишком частые запросы на повторную отправку
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Повторно отправить magic-link
      tags:
      - 2fa
...
swagger: "2.0"